% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cohort_size.R
\name{calculate_cohort_size}
\alias{calculate_cohort_size}
\title{Calculate den cohort size experienced by each cub}
\usage{
calculate_cohort_size(ids)
}
\arguments{
\item{List}{of ids to calculate cohort size for}
}
\value{
A data frame with the ids supplied, size of cohort, whether the id
survived to graduation (1 = yes, 0 = no) and the list of cubs in the cohort.
}
\description{
Calculates a cohort size (and cohort members) experienced by the supplied cub.
Cohorts are specific to each cub. To be included in a cohort, cubs must have a den
period that overlaps with the focal cub by at least 4 months and must be seen
with focal cub at least once while they are both den dependent. Den period is
calculated as (DOB to DenGrad), (DOB to 1yr old) if no DenGrad specified, or
(DOB to Disappeared) if the cub died before graduation or 1 year old.
}
