% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_association_network.R
\name{get_association_network}
\alias{get_association_network}
\title{Get and/or plot hyena association networks}
\usage{
get_association_network(
  clan,
  start,
  end,
  network_duration = "6 months",
  return_nets = F,
  plot_nets = F,
  output_pdf_pathname = NULL
)
}
\arguments{
\item{clan}{One (or more) clan names, supplied as a character vector.}

\item{start}{Start date for desired networks}

\item{end}{End date for desired networks}

\item{network_duration}{Duration of data for each network. If NULL, all data between start and end are lumped into one network.}

\item{return_nets}{Should networks be output?}

\item{plot_nets}{Should networks be plotted?}

\item{If}{\code{plot_nets} is TRUE, path and filename of pdf where plots are output}
}
\value{
Returns weighted, undirected networks for each time period
}
\description{
Creates, plots, and returns association networks for specified clans
}
\examples{
get_association_network(clan = c('serena.s', 'serena.n', 'happy.zebra'), start = '2009-01-01', end = '2025-01-01', network_duration = '12 months', return_nets = F, plot_nets = T, output_pdf_pathname = '~/Desktop/serena_nets.pdf')

}
