% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maternal_relatedness.R
\name{get_maternal_relatedness}
\alias{get_maternal_relatedness}
\title{Get maternal relatedness between two individuals}
\usage{
get_maternal_relatedness(id1, id2)
}
\arguments{
\item{id1}{A string specifying the hyena id for first individual}

\item{id2}{A string specifying the hyena id for second individual}
}
\description{
Returns a value for expected relatedness between two individuals based on maternal pedigree. Returns NA if individuals are unrelated
}
\examples{

}
