% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_descendant.R
\name{is_descendant}
\alias{is_descendant}
\title{Is one hyena a maternal descendant of the other?}
\usage{
is_descendant(test.id, matriarch)
}
\arguments{
\item{test.id}{A string specifying the name of the hypothesized descendant}

\item{matriarch}{A string specifying the name of the hypothesized matriarch}
}
\description{
Returns TRUE or FALSE depending on whether the test.id is a maternal descendant of
the hypothesized matriarch.
}
\examples{
is_descendant('knot', 'kb') # returns TRUE
is_descendant('pike', 'kb') # returns FALSE

}
