% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_clan_membership.R
\name{annual_clan_membership}
\alias{annual_clan_membership}
\title{Identify living hyenas in specified clan on an annual basis}
\usage{
annual_clan_membership(clans, years)
}
\arguments{
\item{years}{One (or more) year, supplied as a numeric vector}

\item{clan}{One (or more) clan names, supplied as a character vector.}
}
\description{
Identifies all hyenas alive in a specified year for the given clan.
}
\examples{
## returns all hyenas from serena clans 2010-2012
annual_clan_membership(clans = c('happy.zebra', 'serena.s', 'serena.n'), years = 2010:2012)

}
