% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maternal_ancestors.R
\name{get_maternal_ancestors}
\alias{get_maternal_ancestors}
\title{Get list of maternal ancestors}
\usage{
get_maternal_ancestors(id)
}
\arguments{
\item{id}{A string specifying the hyena id}
}
\description{
Returns a vector starting with id and ending with most distant maternal ancestor. This function uses tblHyenas
}
\examples{

}
